# api/swagger.yml
...
consumes:
- application/json
produces:
- application/json
swagger: "2.0"
info:
  description: API documentation
  title: test-project
  version: 0.1.0
paths:
...
  /api/v1/auth/forgot-password:
    post:
      description: |-
        Initiates a password reset for a local user, sending an email with a password
        reset link to the provided email address if a user account exists. Will always
        succeed, even if no user was found in order to prevent user enumeration
      tags:
      - auth
      summary: Initiate password reset for local user
      operationId: PostForgotPasswordRoute
      parameters:
      - name: Payload
        in: body
        schema:
          $ref: '#/definitions/postForgotPasswordPayload'
      responses:
        "204":
          description: Success
        "400":
          description: PublicHTTPValidationError
          schema:
            $ref: '#/definitions/publicHttpValidationError'
  /api/v1/auth/forgot-password/complete:
    post:
      description: |-
        Completes a password reset for a local user, using the password reset token sent via email
        to confirm user access, setting the new password if successful. All current access and refresh
        tokens are invalidated and a new set of auth tokens is returned
      tags:
      - auth
      summary: Completes password reset for local user
      operationId: PostForgotPasswordCompleteRoute
      parameters:
      - name: Payload
        in: body
        schema:
          $ref: '#/definitions/postForgotPasswordCompletePayload'
      responses:
        "200":
          description: PostLoginResponse
          schema:
            $ref: '#/definitions/postLoginResponse'
        "400":
          description: PublicHTTPValidationError, type `INVALID_PASSWORD`
          schema:
            $ref: '#/definitions/publicHttpValidationError'
        "403":
          description: PublicHTTPError, type `USER_DEACTIVATED`/`NOT_LOCAL_USER`
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: PublicHTTPError, type `TOKEN_NOT_FOUND`
          schema:
            $ref: '#/definitions/publicHttpError'
        "409":
          description: PublicHTTPError, type `TOKEN_EXPIRED`
          schema:
            $ref: '#/definitions/publicHttpError'
  /api/v1/auth/login:
    post:
      description: Returns an access and refresh token on successful authentication
      tags:
      - auth
      summary: Login with local user
      operationId: PostLoginRoute
      parameters:
      - name: Payload
        in: body
        schema:
          $ref: '#/definitions/postLoginPayload'
      responses:
        "200":
          description: PostLoginResponse
          schema:
            $ref: '#/definitions/postLoginResponse'
        "400":
          description: PublicHTTPValidationError
          schema:
            $ref: '#/definitions/publicHttpValidationError'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "403":
          description: PublicHTTPError, type `USER_DEACTIVATED`
          schema:
            $ref: '#/definitions/publicHttpError'
...
  /api/v1/auth/refresh:
    post:
      description: |-
        Returns a fresh set of access and refresh tokens if a valid refresh token was provided.
        The old refresh token used to authenticate the request will be invalidated.
      tags:
      - auth
      summary: Refresh tokens
      operationId: PostRefreshRoute
      parameters:
      - name: Payload
        in: body
        schema:
          $ref: '#/definitions/postRefreshPayload'
      responses:
        "200":
          description: PostLoginResponse
          schema:
            $ref: '#/definitions/postLoginResponse'
        "400":
          description: PublicHTTPValidationError
          schema:
            $ref: '#/definitions/publicHttpValidationError'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "403":
          description: PublicHTTPError, type `USER_DEACTIVATED`
          schema:
            $ref: '#/definitions/publicHttpError'
  /api/v1/auth/register:
    post:
      description: Returns an access and refresh token on successful registration
      tags:
      - auth
      summary: Registers a local user
      operationId: PostRegisterRoute
      parameters:
      - name: Payload
        in: body
        schema:
          $ref: '#/definitions/postRegisterPayload'
      responses:
        "200":
          description: PostLoginResponse
          schema:
            $ref: '#/definitions/postLoginResponse'
        "400":
          description: PublicHTTPValidationError, type `INVALID_PASSWORD`
          schema:
            $ref: '#/definitions/publicHttpValidationError'
        "409":
          description: PublicHTTPError, type `USER_ALREADY_EXISTS`
          schema:
            $ref: '#/definitions/publicHttpError'
...
  /api/v1/books:
    get:
      security:
      - Bearer: []
      tags:
      - books
      summary: Fetches all the books from the user.
      operationId: getUserBooks
      responses:
        "200":
          description: GetUserBooksResponse
          schema:
            $ref: '#/definitions/getUserBooksResponse'
        "400":
          description: Bad Request
    post:
      security:
      - Bearer: []
      tags:
      - books
      summary: Creates a book.
      operationId: createBook
      parameters:
      - description: book to create
        name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/createBookRequest'
      responses:
        "201":
          description: Added book to database
        "400":
          description: Invalid input
...
  /api/v1/lists:
    get:
      security:
      - Bearer: []
      tags:
      - lists
      summary: Fetches all lists from the user.
      operationId: getAllLists
      responses:
        "200":
          description: Successful operation.
          schema:
            $ref: '#/definitions/getAllListsResponse'
    post:
      security:
      - Bearer: []
      description: Creates a new book list with a name.
      tags:
      - lists
      summary: Creates a new book list.
      operationId: createList
      parameters:
      - description: Creates a new book list with a name.
        name: newList
        in: body
        required: true
        schema:
          $ref: '#/definitions/listRequest'
      responses:
        "201":
          description: List has been created.
        "400":
          description: Invalid input
...
definitions:
  bookInMyDb:
    type: object
    required:
    - title
    - author
    properties:
      author:
        type: string
        example: J. K. Rowling
      book_description:
        type: string
        example: Dark times have come to Hogwarts. After the Dementors' attack on
          his cousin Dudley, Harry Potter knows that Voldemort will stop at nothing
          to find him. ...
      book_id:
        type: string
        format: uuid4
        example: d6764ee3-bf09-40c3-97c5-8f78b7de7ec3
      genre:
        type: string
        example: Novel, Fantasy Fiction, Young adult literature
      pages:
        type: integer
        format: int32
        example: 409
      publisher:
        type: string
        example: Bloomsbury
      rating:
        description: Ratings are from 1 to 5.
        type: integer
        format: int32
        example: 3
      title:
        type: string
        example: Harry Potter
      user_notes:
        type: string
        example: Amazing Character Development from the main character.
  createBookRequest:
    type: object
    required:
    - google_book_ID
    - title
    - author
    properties:
      author:
        type: string
        example: J. K. Rowling
      book_description:
        type: string
        example: Dark times have come to Hogwarts. After the Dementors' attack on
          his cousin Dudley, Harry Potter knows that Voldemort will stop at nothing
          to find him. ...
      genre:
        type: string
        example: Novel, Fantasy Fiction, Young adult literature
      google_book_ID:
        type: string
        example: P3LFEAAAQBAJ
      pages:
        type: integer
        format: int32
        example: 409
      publisher:
        type: string
        example: Bloomsbury
      title:
        type: string
        example: Harry Potter
  getAllListsResponse:
    type: object
    required:
    - data
    properties:
      data:
        description: lists of all lists of user
        type: array
        items:
          $ref: '#/definitions/list'
...
  getUserBooksResponse:
    type: object
    required:
    - data
    properties:
      data:
        description: list of books of user
        type: array
        items:
          $ref: '#/definitions/bookInMyDb'
      pagination:
        $ref: '#/definitions/pagination'
...
  httpValidationErrorDetail:
    type: object
    required:
    - key
    - in
    - error
    properties:
      error:
        description: Error describing field validation failure
        type: string
      in:
        description: Indicates how the invalid field was provided
        type: string
      key:
        description: Key of field failing validation
        type: string
  list:
    type: object
    required:
    - list_id
    - name
    - user_id
    properties:
      books:
        type: array
        items:
          $ref: '#/definitions/bookInMyDb'
      list_id:
        type: string
        format: uuid
        example: d6764ee3-bf09-40c3-97c5-8f78b7de7ec3
      name:
        type: string
        example: Cookbooks
      user_id:
        type: string
        format: uuid
        example: d6764ee3-bf09-40c3-97c5-8f78b7de7ec3
  listRequest:
    type: object
    required:
    - name
    properties:
      name:
        type: string
        example: Favorites
...
  pagination:
    type: object
    properties:
      currentPage:
        description: Current page the user is on
        type: integer
      hasNextPage:
        description: Is there another page after?
        type: boolean
      hasPreviousPage:
        description: Is there a page before?
        type: boolean
      pageSize:
        description: Amount of elements per page
        type: integer
      totalItems:
        description: Sum of all existing items
        type: integer
      totalPages:
        description: Sum of all pages
        type: integer
...
  postForgotPasswordCompletePayload:
    type: object
    required:
    - token
    - password
    properties:
      password:
        description: New password to set for user
        type: string
        maxLength: 500
        minLength: 1
        example: correct horse battery staple
      token:
        description: Password reset token sent via email
        type: string
        format: uuid4
        example: ec16f032-3c44-4148-bbcc-45557466fa74
  postForgotPasswordPayload:
    type: object
    required:
    - username
    properties:
      username:
        description: Username to initiate password reset for
        type: string
        format: email
        maxLength: 255
        minLength: 1
        example: user@example.com
  postLoginPayload:
    type: object
    required:
    - username
    - password
    properties:
      password:
        description: Password of user to authenticate as
        type: string
        maxLength: 500
        minLength: 1
        example: correct horse battery staple
      username:
        description: Username of user to authenticate as
        type: string
        format: email
        maxLength: 255
        minLength: 1
        example: user@example.com
  postLoginResponse:
    type: object
    required:
    - access_token
    - token_type
    - expires_in
    - refresh_token
    properties:
      access_token:
        description: Access token required for accessing protected API endpoints
        type: string
        format: uuid4
        example: c1247d8d-0d65-41c4-bc86-ec041d2ac437
      expires_in:
        description: Access token expiry in seconds
        type: integer
        format: int64
        example: 86400
      refresh_token:
        description: Refresh token for refreshing the access token once it expires
        type: string
        format: uuid4
        example: 1dadb3bd-50d8-485d-83a3-6111392568f0
      token_type:
        description: Type of access token, will always be `bearer`
        type: string
        example: bearer
...
  postRefreshPayload:
    type: object
    required:
    - refresh_token
    properties:
      refresh_token:
        description: Refresh token to use for retrieving new token set
        type: string
        format: uuid4
        example: 7503cd8a-c921-4368-a32d-6c1d01d86da9
  postRegisterPayload:
    type: object
    required:
    - username
    - password
    properties:
      password:
        description: Password to register with
        type: string
        maxLength: 500
        minLength: 1
        example: correct horse battery staple
      username:
        description: Username to register with
        type: string
        format: email
        maxLength: 255
        minLength: 1
        example: user@example.com
...
  publicHttpError:
    type: object
    required:
    - status
    - type
    - title
    properties:
      detail:
        description: More detailed, human-readable, optional explanation of the error
        type: string
        example: User is lacking permission to access this resource
      status:
        description: HTTP status code returned for the error
        type: integer
        format: int64
        maximum: 599
        minimum: 100
        x-go-name: Code
        example: 403
      title:
        description: Short, human-readable description of the error
        type: string
        example: Forbidden
      type:
        description: Type of error returned, should be used for client-side error
          handling
        type: string
        example: generic
  publicHttpValidationError:
    type: object
    required:
    - validationErrors
    allOf:
    - $ref: '#/definitions/publicHttpError'
    properties:
      validationErrors:
        description: List of errors received while validating payload against schema
        type: array
        items:
          $ref: '#/definitions/httpValidationErrorDetail'
...
securityDefinitions:
  Bearer:
    description: |-
      Access token for application access, **must** include "Bearer " prefix.
      Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
    type: apiKey
    name: Authorization
    in: header
    x-keyPrefix: 'Bearer '
  Management:
    description: Management secret, used for monitoring and infrastructure related
      calls
    type: apiKey
    name: mgmt-secret
    in: query